fix(datagrid): omit Pagination for small client-side paginated collections (#461)
Two changes:

- `minrows` prop set on `DataGrid` (so we don't have lots of empty rows)
- `Pagination` conditionally renders

The `Pagination` component conditionally returns `null` if 1) the data paged client-side, and 2) if the number of pages (`instance#pages`, -computed by collection size and page size) is less than or equal to 1. The no-scrollbar-x styles used for `Breadcrumb`s was moved to a sass mixin and added to `DataGrid`'s `-tbody` class. The guide site docs were updated to include collections of sufficiently small size.