chore(build): PR pipeline, docker beginnings (#463)
Adds a PR build Jenkins pipeline. Uses the [Docker Workflow plugin](https://go.cloudbees.com/docs/plugins/docker-workflow/)'s `.inside()` to execute cmds from within the container, so we don't need a Dockerfile or need to bake an image. Left the `stage helper` fns defined, though we don't use them in this pipeline. Sets the `--user` arg in `docker run` so that yarn cache may be created and be writable.