add moment, simple docs, notes in Guide
Its unclear if react-widgets#DateTimePicker will be our widget for this
need in the future... Its nice that there's scss exported in the dist
from `react-widgets`, but the cognitive load of having to manage
_multiple_ sass-based systems is taxing. Plus, there's no good way I can
surmise having used the lib for a day how to pass components into the
renders... (e.g.; pass our `Button` to `DateTimePicker` so that that we
are COMPOSING things from our primitives).

Probably stopping here CL-1 here until we can draw some conclusions.