WiP - (notes below)
Ran into an issue using a simple boolMap (`{ [key: string]: boolean }`)
for checkbox value tracking. A desired feature of `@cwds/forms` is the
interoperability of FormComponents with DataTypes... (e.g.; our
`N-pick-k` -type collections should be renderable as either
`Select[multi]`, `CheckboxBank`, etc. by simply chaning the render
component.) The issue is that `Select[multi]` needs _order_ to render
from values because it arranges in the `chips` in the order you add
them (a stack).

So let's try something else...

- Use an array datastructure for value tracking
- Use the `FieldArray` instead instead of `Field`